refactor: 로컬스토리지 카드 등록 로직 분리
